+++ packages/core/tests/Unit/Models/CartTest.php
@@ -8,6 +8,7 @@ use Lunar\DataTypes\Price as DataTypesPrice;
 use Lunar\DataTypes\ShippingOption;
 use Lunar\DiscountTypes\AmountOff;
 use Lunar\Exceptions\Carts\CartException;
+use Lunar\Exceptions\FingerprintMismatchException;
 use Lunar\Facades\Discounts;
 use Lunar\Facades\ShippingManifest;
 use Lunar\Models\Cart;
@@ -139,7 +140,7 @@ class CartTest extends TestCase
             'user_id' => $user->getKey(),
         ]);
 
-        $this->assertDatabaseHas((new Cart)->getTable(), [
+        $this->assertDatabaseHas((new Cart())->getTable(), [
             'currency_id' => $currency->id,
             'channel_id' => $channel->id,
             'user_id' => $user->getKey(),
@@ -296,7 +297,7 @@ class CartTest extends TestCase
             'user_id' => $user->getKey(),
         ]);
 
-        $this->assertDatabaseHas((new Cart)->getTable(), [
+        $this->assertDatabaseHas((new Cart())->getTable(), [
             'currency_id' => $currency->id,
             'channel_id' => $channel->id,
             'user_id' => $user->getKey(),
@@ -653,5 +654,52 @@ class CartTest extends TestCase
 
         $this->assertCount(1, $cart->discountBreakdown);
         $this->assertSame(10, $cart->discountBreakdown->first()->price->value);
+    }
+
+    /** @test */
+    public function can_validate_fingerprint()
+    {
+        $currency = Currency::factory()->create();
+        $channel = Channel::factory()->create();
+
+        $cart = Cart::create([
+            'currency_id' => $currency->id,
+            'channel_id' => $channel->id,
+            'meta' => [
+                'A' => 'B',
+                'C' => 'D',
+            ],
+        ]);
+
+        $variant = ProductVariant::factory()->create();
+
+        Price::factory()->create([
+            'price' => 100,
+            'tier' => 1,
+            'currency_id' => $currency->id,
+            'priceable_type' => get_class($variant),
+            'priceable_id' => $variant->id,
+        ]);
+
+        $cart->lines()->create([
+            'purchasable_type' => ProductVariant::class,
+            'purchasable_id' => $variant->id,
+            'quantity' => 1,
+        ]);
+
+
+        $fingerprint = $cart->fingerprint();
+
+        $this->assertTrue(
+            $cart->checkFingerprint($fingerprint)
+        );
+
+        $cart->update([
+            'coupon_code' => 'FOOBAR',
+        ]);
+
+        $this->expectException(FingerprintMismatchException::class);
+
+        $cart->checkFingerprint($fingerprint);
     }
 }